+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 ------------------
 
 - #2654 Show Batch title or ID in Sample reference field
+- #2657 Methods from analyses are not updated on instrument change in worksheet
+- #2656 Fix AnalysisProfile keyword validator fail with non-ascii value 
+- #2646 Add SelectOtherField and SelectOtherWidget
 - #2649 Fix event subscribers are not triggered on analysis initialization
 - #2650 Fix AttributeError on analysis update with interim values via jsonapi
 - #2648 Increase the default width for field labels to min 150px